feat(utils): Expose `calculateContrast`, `calculateLuminance` and `hexToRGB` utility functions
