+++ .changeset/chilled-cobras-develop.md
@@ -1,0 +1,5 @@
+---
+'@noaignite/utils': minor
+---
+
+Expose `calculateContrast`, `calculateLuminance` and `hexToRGB` utility functions
