+++ packages/utils/src/colorContrast.ts
@@ -1,46 +1,6 @@
-/**
- * Helper function to calculate relative luminance for a color
- */
-const calculateLuminance = (rgb: [number, number, number]): number => {
-  const [r, g, b] = rgb.map((c) => {
-    const sRGB = c / 255
-    return sRGB <= 0.03928 ? sRGB / 12.92 : ((sRGB + 0.055) / 1.055) ** 2.4
-  }) as [number, number, number]
-
-  const luminance = 0.2126 * r + 0.7152 * g + 0.0722 * b
-  return luminance
-}
-
-/**
- * Helper function to calculate contrast ratio between two colors
- */
-const calculateContrast = (rgb1: [number, number, number], rgb2: [number, number, number]) => {
-  const luminance1 = calculateLuminance(rgb1)
-  const luminance2 = calculateLuminance(rgb2)
-  const contrast =
-    (Math.max(luminance1, luminance2) + 0.05) / (Math.min(luminance1, luminance2) + 0.05)
-  return contrast
-}
-
-/**
- * Helper function to convert hex to RGB
- */
-const hexToRGB = (hex: string): [number, number, number] => {
-  let normalizedHex = hex.replace(/^#/, '')
-
-  // Expand shorthand hex notation
-  if (normalizedHex.length === 3) {
-    normalizedHex = normalizedHex
-      .split('')
-      .map((char) => char + char)
-      .join('')
-  }
-
-  const r = parseInt(normalizedHex.substring(0, 2), 16)
-  const g = parseInt(normalizedHex.substring(2, 4), 16)
-  const b = parseInt(normalizedHex.substring(4, 6), 16)
-  return [r, g, b]
-}
+import { calculateContrast } from './calculateContrast'
+import { calculateLuminance } from './calculateLuminance'
+import { hexToRGB } from './hexToRGB'
 
 /**
  * Accepts a HEX `baseColor` and any number of HEX `colors` and returns the
