+++ packages/utils/src/index.ts
@@ -1,8 +1,11 @@
 export * from './assert'
+export * from './calculateContrast'
+export * from './calculateLuminance'
 export * from './capitalize'
 export * from './clamp'
 export * from './colorContrast'
 export * from './deepmerge'
+export * from './hexToRGB'
 export * from './isObject'
 export * from './isPlainObject'
 export * from './lerp'
